Export to OrientDB(splitted "exportOrientDbFunctions" to separate classes with unit tests)
