+++ engine-plugins/graph-plugins/src/main/scala/org/trustedanalytics/atk/plugins/orientdb/ExportOrientDbGraphPlugin.scala
@@ -1,17 +1,17 @@
 /**
- * Copyright (c) 2015 Intel Corporation 
+ *  Copyright (c) 2015 Intel Corporation 
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
  *       http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 package org.trustedanalytics.atk.plugins.orientdb
 
@@ -82,7 +82,8 @@ class ExportOrientDbGraphPlugin extends SparkCommandPlugin[ExportOrientDbGraphAr
       //load frame as RDD
       val sparkFrame: SparkFrame = frame
       val vertexFrameRdd = new VertexFrameRdd(sparkFrame.rdd)
-      val verticesCount = ExportOrientDbFunctions.exportVertexFrame(arguments.dbUri, vertexFrameRdd, arguments.batchSize)
+      val exportVertexFrame = new VertexFrameWriter
+      val verticesCount = exportVertexFrame.exportVertexFrame(arguments.dbUri, vertexFrameRdd, arguments.batchSize)
 
       (vertexFrameRdd.vertexWrapper.schema.label, verticesCount)
     })
@@ -105,7 +106,8 @@ class ExportOrientDbGraphPlugin extends SparkCommandPlugin[ExportOrientDbGraphAr
       //load frame as RDD
       val sparkFrame: SparkFrame = frame
       val edgeFrameRdd = new EdgeFrameRdd(sparkFrame.rdd)
-      val edgesCount = ExportOrientDbFunctions.exportEdgeFrame(arguments.dbUri, edgeFrameRdd, arguments.batchSize)
+      val exportEdgeFrame = new EdgeFrameWriter
+      val edgesCount = exportEdgeFrame.exportEdgeFrame(arguments.dbUri, edgeFrameRdd, arguments.batchSize)
       (edgeFrameRdd.edge.schema.label, edgesCount)
     })
     metadata.toMap
