+++ engine-plugins/graph-plugins/src/main/scala/org/trustedanalytics/atk/plugins/orientdb/GraphDbFactory.scala
@@ -1,3 +1,18 @@
+/**
+ *  Copyright (c) 2015 Intel Corporation 
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package org.trustedanalytics.atk.plugins.orientdb
 
 import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx
@@ -7,16 +22,16 @@ import com.tinkerpop.blueprints.impls.orient.{ OrientGraphFactory, OrientGraph }
  * Created by wtaie on 4/1/16.
  */
 
-object GraphDbFactory {
-
+class GraphDbFactory {
   /**
    * Method to create/connect to Orient database
    * @param dbUri the database with default username; "admin" and password : "admin".
    * @return a transcational Orient graph database instance
    */
-  def GraphDbConnector(dbUri: String): OrientGraph = {
-    val orientDb: ODatabaseDocumentTx = new ODatabaseDocumentTx(dbUri)
 
+  def GraphDbConnector(dbUri: String): OrientGraph = {
+
+    val orientDb: ODatabaseDocumentTx = new ODatabaseDocumentTx(dbUri)
     if (!orientDb.exists()) {
       orientDb.create()
     }
@@ -36,6 +51,7 @@ object GraphDbFactory {
    * @return a transcational Orient graph database instance
    */
   def createGraphDb(dbUri: String, userName: String, password: String): OrientGraph = {
+
     val factory: OrientGraphFactory = new OrientGraphFactory(dbUri, userName, password)
     var graph: OrientGraph = factory.getTx
     graph
