+++ engine-plugins/graph-plugins/src/test/scala/org/trustedanalytics/atk/plugins/orientdb/VertexFrameWriterTest.scala
@@ -1,16 +1,31 @@
+/**
+ *  Copyright (c) 2015 Intel Corporation 
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package org.trustedanalytics.atk.plugins.orientdb
 
 import org.apache.spark.atk.graph.VertexFrameRdd
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.expressions.GenericRow
-import org.scalatest.{ WordSpec, BeforeAndAfterEach, Matchers }
-import org.trustedanalytics.atk.domain.schema._
+import org.scalatest.{ BeforeAndAfterEach, Matchers, WordSpec }
+import org.trustedanalytics.atk.domain.schema.{ VertexSchema, DataTypes, GraphSchema, Column }
 import org.trustedanalytics.atk.testutils.{ TestingOrientDb, TestingSparkContextWordSpec }
 
 /**
- * Created by wtaie on 4/5/16.
+ * Created by wtaie on 4/18/16.
  */
-class ExportOrientDbFunctionsVertexFrameTest extends WordSpec with TestingSparkContextWordSpec with Matchers with TestingOrientDb with BeforeAndAfterEach {
+class VertexFrameWriterTest extends WordSpec with TestingSparkContextWordSpec with Matchers with TestingOrientDb with BeforeAndAfterEach {
   override def beforeEach() {
     setupOrientDb()
   }
@@ -18,9 +33,8 @@ class ExportOrientDbFunctionsVertexFrameTest extends WordSpec with TestingSparkC
   override def afterEach() {
     cleanupOrientDb()
   }
-
-  "Export orientDb functions" should {
-    "test export vertex frame" in {
+  "vertex frame writer" should {
+    "export vertex frame to OrientDB" in {
       val dbUri: String = "plocal:/home/wtaie/graphDBs_home/orientdb-community-2.1.12/databases/OrientDbTest"
       val columns = List(Column(GraphSchema.vidProperty, DataTypes.int64), Column(GraphSchema.labelProperty, DataTypes.string), Column("name", DataTypes.string), Column("from", DataTypes.string), Column("to", DataTypes.string), Column("fair", DataTypes.int32))
       val schema = new VertexSchema(columns, GraphSchema.labelProperty, null)
@@ -34,7 +48,8 @@ class ExportOrientDbFunctionsVertexFrameTest extends WordSpec with TestingSparkC
       val batchSize = 4
       val rowRdd = sparkContext.parallelize(vertices)
       val vertexFrameRdd = new VertexFrameRdd(schema, rowRdd)
-      val verticesCount = ExportOrientDbFunctions.exportVertexFrame(dbUri, vertexFrameRdd, batchSize)
+      val vertexFrameWriter = new VertexFrameWriter
+      val verticesCount = vertexFrameWriter.exportVertexFrame(dbUri, vertexFrameRdd, batchSize)
       verticesCount shouldEqual (4)
     }
   }
